Kotlin. ДЗ 1, Шеметова Екатерина
Task 51B
http://codeforces.com/contest/51/submission/42588052
