--- kotlin-course-se-01-just-for-fun/README.md
@@ -1,34 +1,0 @@
-# ДЗ01 - Just for fun
-
-**Дедлайн**: 2018.09.12 23:59 (полный балл), 2018.09.19 23:59 (50%)
-
-Необходимо решить задачу с сайта [http://codeforces.com/](codeforces.com),
-используя в качестве языка решения Kotlin.
-
-- Задача должна быть не проще C из второго дивизиона (или можете выбрать любую из списка ниже).
-- Будет здорово, если тематика задачи будет как-то связана с графами/геометрией/структурами
-данных/строками. Короче с чем-то, где кода будет чуть больше, чем одна формула.
-- Требования к качеству кода при этом будут строгие, как к исходникам какого-нибудь настоящего проекта.
-- В PR с решением должен быть файл main.kt, который должно быть можно отправить в качестве решения,
-и unit-тесты, построенные на основе тестов из примеров.
-- В комментарии к PR должна быть ссылка на успешный сабмит
-- Если вам кажется, что давать олимпиадные задачи на курс про язык программирования неправильно,
-то это очень грустно. Однако, здесь не запрещено смотреть в код
-других решений на сайте (не на Kotlin и желательно не на Java).
-- Ваше решение должно быть можно собрать с помощью Gradle. Можно использовать шаблон, который находится в данной ветке.
-
-Варианты задач:
-- http://codeforces.com/problemset/problem/120/J
-- http://codeforces.com/problemset/problem/701/E
-- http://codeforces.com/problemset/problem/590/B
-- http://codeforces.com/problemset/problem/159/C
-- http://codeforces.com/problemset/problem/788/B
-- http://codeforces.com/problemset/problem/598/C
-- http://codeforces.com/problemset/problem/59/C
-- http://codeforces.com/problemset/problem/610/D
-- http://codeforces.com/problemset/problem/849/D
-- http://codeforces.com/contest/854/problem/C
-- http://codeforces.com/problemset/problem/131/D
-- http://codeforces.com/problemset/problem/200/D
-- http://codeforces.com/problemset/problem/51/B
-- http://codeforces.com/problemset/problem/113/A
