--- kotlin-course-se-01-just-for-fun/settings.gradle
@@ -1,11 +1,0 @@
-pluginManagement {
-    repositories {
-        mavenCentral()
-
-        maven {
-            url 'https://plugins.gradle.org/m2/'
-        }
-    }
-}
-rootProject.name = '01-just-for-fun'
-
